Memory-Adaptation-In-Stimulus-drift
**Memory-Adaptation-In-Stimulus-Drift**
This repository hosts a cognitive experiment implemented as a web-based form, designed to collect human data for studying memory adaptation in response to stimulus drift. The experiment is used in our lab to gather essential data for exploring how individuals adjust their memory and decision-making to gradual changes in stimuli over time. The repository includes all necessary code and resources to run the experiment and record participant responses.